extract sandbox badge component
